backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
stripe==11.2.0
celery==5.4.0
redis==5.2.0
anthropic==0.40.0
tenacity==9.1.4
python-dotenv==1.2.2
alembic==1.14.0
